rename s3 to object storage
